Calendar Resource code refactorization + support for creating recurring events
This commit makes some code refactorization of Calendar Resource's code.
The new scheme is similar to that in Contacts Resource - e.g. Event::Event class
was introduced that povides API for conversion to/from JSON.

Additionally, creating recurring events was fixed so that Google now accepts
such events as valid.