Fix parsers to correctly parse complete documents
XML and JSON parser must be able to parse not only body of a document
(data representing the contact/event itself) but the entire document
as well (including the header, etc), respectively they must be able to find the
body within the document and then parse it.

This fixes the issue that newly created objects didn't receive remoteId and
remoteRevision - the values were not correctly parsed from server response and
thus were not assigned to the item.