Wait for all events to be fetched before calling itemsRetrieved() (fixes bug#282509)
Items in Calendar resource are queued in a list until GAM emits
requestFinished() signal (when last page of feed is retrieved) and
then commits the whole list at once (using itemsRetrieved() or
itemsRetrievedIncremental()).

If this appears to work as expected, it should be implemented in other
resources as well to prevent crashes like in bug #282509 and #282799.