Introduce KGoogle::Account class and common.h
This class represents a single Google account, holds authentication tokens
and list of scopes the tokens are valid for.

Introduce common.h as header with common values and objects. For now it
contains KGoogle::Error enumarator and few basic exceptions

The "KGoogle" in name was dropped since all classes are in KGoogle namespace
so there is no need to have KGoogle::KGoogleSomeClass.