Refactor and rewrite KGoogleAuth and AuthDialog.
KGoogleAuth is now KGoogle::Auth (in auth.h and auth.cpp). It
supports multiple Google accounts and acts as a frontend to a
secure storage (KWallet in this case). It still provides methods
for authenticating new accounts or refreshing access tokens to
an existing one, the API is much more simple and straightforward though.