Minimize access to KWallet by storing Account in resource
Instead of asking KGoogle::Auth (and thus KWallet) for Account every time
resource wants to send a request, the Account is stored in resource
configuration.
Since the Account::Ptr is an explicitely shared pointer, updated tokens
will be updated in all it's copies, thus in the copy stored in the
resources as well.