Use internal cache for Accounts in KGoogle::Auth
When the item is stored in cache, KGoogle::Auth can update it globally, since
Account is explicitely shared. This means, that updating the Account from one
place in an application will update it in all it's copies in the application.

KGoogle::Auth also watches for changes in KWallet and can update the Account
when the data in Kwallet changes, for example when tokens are updated by a
different application.