Don't update event UID after it's created
Untill now we tried to update UID of a newly created event after we received
the UID from Google, but KOrganizer didn't cope well with this.

This patch removes the IteModifyJob which was updating the UID. We rather
keep the KOrganizer-generated UID and we only change it in the payload before
serializing it and sending it to Google (for instance when we want to modify
the event). This makes KOrganizer happy.

BUG: 298518
FIXED-IN: 0.3.1