Fix a crash by cleaning the ongoing Fetch jobs when quitting the resource
Many methods within ResourceBase check if the current task is of a type
that makes sense, for example if a method for synchronitation is called
the current task should be of type Sync. Problem is that they check
this in an ASSERT and we are hitting it when resource is Quit.

What this patch does is track the fetchJobs and clean them in
aboutToQuit.

REVIEW:104419