Handle Microsoft TZIDs
Microsoft TZIDs which are not compatible with the Olson timezone
format which Google expects in dtStart and dtEnd timezones. We
use special mapping of X-MICROSOFT-CDO-TZID property to a real
Olson timezones to make Google happy. Unfortunatelly this means
that the original long MS timezone name is not preserved.

The conversion is run only in Objects::Event -> JSON serializer,
as Google always returns events with Olson timezone format.

BUG: 310851
FIXED-IN: 0.4.4