LibKGAPI 2 API (breaks ABI, the old API is preserved)
LibKGAPI2 is a new job-based API that deprecates the
old one. It is cleaner, more flexible and much easier
to use.

The old API is still available, but all services
internally call services from LibKGAPI2. Implementation
of KGAPI::Objects::* classes is also shared with the
one from LibKGAPI2 to avoid as much code duplication
as possible. It also means that fixing serializer bug
in LibKGAPI2 will fix it in the old API as well
(although the old API is not officially supported as
of now). Thanks to this, all distributions can safely
update to LibKGAPI-2.0 without breaking any
applications (you just need to rebuild them, because I
had to break ABI).

The old API will be removed probably after KDE 4.12 is
released.

As of now master is pointing to LibKGAPI-2.0 and there
will be no more updates in LibKGAPI/0.4 branch.