Don't skip item when restarting a job
When a recoverable error occured (invalid tokens), the job
skipped the item that has failed and tried to continued with
the next one after being restarted. With single-item
jobs this caused clients that expected exactly one item in reply
to crash.

With QueueHelper, the job will re-submit the failed item after
restart before cnotinuing with the next in queue.