When fetching last contact photo fails, don't fail the entire job
Job::replyReceived() sets error to NotFound if it receives 404. In case
of ContactPhotosFetchJob that's not a fatal error though, but a notification
that the contact does not have a photo. Subsequent success will reset the
error, but in case the very last contact does not have a photo, the error
would remain set to NotFound, which would then fail the entire task in Google
resource.

BUG: 328422
FIXED-IN: 2.0.3