Drop KDELibs4Support, fix CMake dependencies, enable Qt 5 deprecated methods
Get rid of KDELibs4Support completely - we don't really require
it anymore for some time, but it was setting some flags that
make the code compile against Qt 5).

Fix CMake dependencies, use COMPONENTS.

Define -DQT_DISABLE_DEPRECATED_BEFORE=0x049999 to enable everything
that was in Qt 4 but was deprecated in Qt 5.0 - in our case this is
only QUrl::addQueryItem. Once we port to QUrlQuery, we can drop this.