Fix URL for simple upload without metadata
Similar to commit 8ac58198f9, we need the /upload prefix in the URL of upload
requests without metadata (`uploadType=media`).
This fixes applications that edit and save a gdrive:/ URL (e.g. Kate with a
text file).

See https://developers.google.com/drive/v2/web/manage-uploads as reference.

Note that the /upload prefix in DriveService::uploadMetadataFileUrl() breaks
at least the creation of empty folders in kio-gdrive, so those URLs are fine as
they are.