Rename CMake target prefix from KF5 to KPim
LibKGAPI is not a Framework (yet), but part of KDE PIM, so use
the KPim:: target prefix. That means that KF5::GAPICore is now
KPim::GAPICore etc. The KF5:: aliases as well as KF5GAPI CMake
config files are still present for backwards compatibility, but
will be eventually removed. When using find_package(KF5GAPI),
a deprecation warning is displayed with instructions to switch
to KPimGAPI.

This is in preparation of moving KGAPI into KDE Applications.