Retrieve OAuth2 token with HTTP socket
Summary:
The webview gets redirected to a local socket and sends a http request to it.
The requests argument includes the token. Thus, we don't need to guess the
token page URL (which led to problems in the past).
The old method is deprecated and this method is the recommended one.

Test Plan:
- Login from KOrganizer and KMail works for my two accounts from two PCs
  (verified with debug logs and a working synchronization)
- I did not test the error case

Differential Revision: https://phabricator.kde.org/D9885