Check for the correct "kind" strings
LibKGAPI does a sanity check on the data that was received
from the Google API to make sure it contains the expected kind of resources.

In src/drive/parentreference.cpp the "kind" string that the code looks for
was "kind#parentList" rather than "drive#parentList" as actually returned by the Google API.

Therefore, further data processing was aborted and the class would always
return an empty ParentReferencesList result.
After correcting this I was able to fetch the desired non-empty output.

src/drive/app.cpp was affected by the same problem.

References:
https://phabricator.kde.org/T3443
https://developers.google.com/drive/v2/reference/parents/list
https://developers.google.com/drive/v2/reference/apps

Differential Revision: https://phabricator.kde.org/D9774