Google tasks are all-day tasks
In the current Google tasks api
https://developers.google.com/tasks/reference/rest/v1/tasks#Task
the "due" field only contains date information;  the time portion is fixed
at 00:00 UTC.  Treating it as a meaningful UTC date-time causes KOrganizer
to display the task at an arbitrary time, and probably on the wrong day.
Treating Google tasks as all-day tasks works around the problem.