First attempt to fix bug 328616
As reported in https://bugs.kde.org/show_bug.cgi?id=328616, when a user
accepts an invitation to an event, when it is loaded on Google Calendar,
it creates a new event with the user as the organizer. To prevent this,
the "import" API should be used instead of the "insert" API.
With this change, the organizer is checked against the account name of
the user: if they match, the event is created (previous implementation),
otherwise the event is imported ("import" API).
This still requires some tests to check if it does not break anything.

BUG: 328616
FIXED-IN: 5.17