saslplugin: Fix build under C99 on all platforms
C99 changes the meaning of `extern inline` (https://stackoverflow.com/questions/216510/what-does-extern-inline-do).
In our case that has the effect that some headers from glibc are treated
differently, i.e. functions like `atof` are actually defined and exported
in the object files including the header defining `atof` and friends.

More information about this issue:
  https://gcc.gnu.org/legacy-ml/gcc/2006-11/msg00011.html

Normally this should not be the case; the underlying issue is that libkgapi
is using a generated config.h (./src/saslplugin/config.h) which is causing
issues under C99.

Fix this by generating the config.h file using CMake.

This patch also updates the 3rdparty plugin_common.* to a more recent
version (cyrus-sasl-2.1.27 to be exact), providing also support for
"Windows 10 or greater".

BUG: 441849