Replace _POSIX_SOURCE->_GNU_SOURCE for saslplugin
The issue here is that NI_MAXHOST is used inside
saslplugin/plugin_common.c. That is a GNU extension only defined when
_GNU_SOURCE is defined. Declarning that this is POSIX source code is
therefore misleading (_POSIX_SOURCE).

This replaces
https://invent.kde.org/pim/libkgapi/-/merge_requests/25
https://invent.kde.org/bribbers/libkgapi/-/commit/a11d8f31ccd25fe6766a4d736327a4e2175bf9cb because it's defining the source code as GNU and then imports a header re-defines the source code as non-GNU.
This instead sets _GNU_SOURCE in the build system for salsplugin which obsoletes the previous MR. 

Another solution would be to for example use IPPORT_RESERVED if you want
to stay POSIX-pure.

Signed-off-by: Alfred Persson Forsberg <cat@catcream.org>