Don't cache promises for AccountManager::findAccount
Unlike AccountManager::getAccount and AccountManager::refreshTokens, this
method does not return an authenticated account. However, the promises are
cached for all of them in the same store, so it was possible for a call to
e.g. refreshTokens to get a promise created by findAccount instead, resulting
in an unexpected result. Just don't cache promises created by findAccount.

BUG: 406839
BUG: 409122
BUG: 421664
BUG: 456923