People API: handle EXPIRED_SYNC_TOKEN error properly
The People API returns HTTP 400 with JSON message describing the error.
In order to handle this, we need a bigger change to the overall error
handling in KGAPI2::Job. It is now possible for each job to first try
to handle the error on its own (regardless of the error type) and
only fall back to the default existing error handling code if the job
doesn't handle it.